remove macro.c
